Implemented apps section
